backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pdfplumber==0.11.4
PyMuPDF==1.24.10
Pillow==11.0.0
requests==2.32.3
tiktoken
numpy

# 向量数据库（需要 >= 0.5.0 以兼容 numpy 2.0）
chromadb>=0.5.0

# 进度条
tqdm==4.66.1

# 图谱处理
networkx==3.2.1

# 关键字检索
jieba==0.42.1
rank-bm25==0.2.2
